Added : Autocomplete Update Marker/RecyclerView on Map and Restaurant Fragment. Button shown when autocomplete to remember autocomplete search with restaurant_name as text. By clicking on this button autocomplete is canceled (generation of marker and recyclerview as application launched). Modification of Restaurant choice : user can click on floating button to delete his choice in restaurant. Bug correction of Language Button radio.

+++ app/src/main/java/com/duboscq/nicolas/go4lunch/api/APIGoogleInterface.java
@@ -10,7 +10,7 @@ import retrofit2.http.Query;
 public interface APIGoogleInterface {
 
     @GET("nearbysearch/json?type=restaurant")
-    Observable<RestaurantPlace> getRestaurantList(@Query("radius") int radius,
+    Observable<RestaurantPlace> getRestaurantList(@Query("radius")double radius,
                                                   @Query("key") String key,
                                                   @Query("location") String location);
